requirements.txt
a2a-sdk==0.3.26
absl-py==2.4.0
ag-ui-protocol==0.1.18
agent-client-protocol==0.9.0
agentscope==1.0.19
agentscope-runtime==1.1.4
aiofile==3.9.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiohttp_socks==0.11.0
aioitertools==0.13.0
aiosignal==1.4.0
alibabacloud-credentials==1.0.8
alibabacloud-credentials-api==1.0.0
alibabacloud-dingtalk==2.2.43
alibabacloud-openapi-util==0.2.4
alibabacloud-tea==0.4.3
alibabacloud-tea-openapi==0.4.4
alibabacloud-tea-util==0.3.14
alibabacloud_endpoint_util==0.0.4
alibabacloud_gateway_dingtalk==1.0.2
alibabacloud_gateway_spi==0.0.3
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altgraph==0.17.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.12.1
APScheduler==3.11.2
asgiref==3.11.1
asttokens==3.0.0
astunparse==1.6.3
asyncio==4.0.0
asyncpg==0.31.0
attrs==26.1.0
audioop-lts==0.2.2
Authlib==1.7.0
auto-py-to-exe==2.47.0
Automat==25.4.16
av==17.0.0
backoff==2.2.1
bcrypt==5.0.0
beartype==0.22.9
bidict==0.23.1
billiard==4.2.4
bottle==0.13.4
bottle-websocket==0.2.9
build==1.4.4
cached-property==2.0.1
cachetools==7.0.6
caio==0.9.25
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
clr_loader==0.2.10
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
constantly==23.10.4
contourpy==1.3.3
crcmod==1.7
cryptography==46.0.7
cssselect==1.4.0
cua==0.1.6
cua-agent==0.8.2
cua-auto==0.1.2
cua-cli==0.1.10
cua-computer==0.5.18
cua-core==0.3.1
cua-sandbox==0.1.12
cycler==0.12.1
cyclopts==4.11.0
darabonba-core==1.0.5
dashscope==1.25.17
debugpy==1.8.16
decorator==5.2.1
dingtalk-stream==0.24.3
discord.py==2.7.1
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.18.0
docutils==0.22.4
dotenv==0.9.9
durationpy==0.10
Eel==0.18.2
elastic-transport==9.2.1
elasticsearch==9.3.0
email-validator==2.3.0
et_xmlfile==2.0.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.136.0
fastmcp==3.2.4
fastuuid==0.14.0
filelock==3.28.0
filetype==1.2.0
fire==0.7.1
flatbuffers==25.12.19
fonttools==4.59.2
frozenlist==1.8.0
fsspec==2026.3.0
future==1.0.0
gast==0.7.0
gevent==25.9.1
gevent-websocket==0.10.1
google-api-core==2.30.3
google-auth==2.49.2
google-genai==1.73.1
google-pasta==0.2.0
googleapis-common-protos==1.74.0
greenlet==3.2.4
griffelib==2.0.2
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.10.2
humanfriendly==10.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Incremental==24.11.0
invoke==3.0.3
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
je-open-cv==0.0.22
jedi==0.19.2
Jinja2==3.1.6
jiter==0.14.0
jmespath==0.10.0
joserfc==1.6.4
json5==0.14.0
json_repair==0.59.4
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.13.2
keyring==25.7.0
kiwisolver==1.4.9
kombu==5.6.2
kubernetes==35.0.0
lark-oapi==1.5.5
libclang==18.1.1
litellm==1.80.0
loguru==0.7.3
lxml==6.0.4
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
matrix-nio==0.25.2
mcp==1.27.0
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.1
modelscope==1.36.1
more-itertools==11.0.2
mpmath==1.3.0
mslex==1.3.0
mss==10.1.0
multidict==6.7.1
namex==0.1.0
nest-asyncio==1.6.0
Nuitka==2.7.16
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.32.0
openapi-pydantic==0.5.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-exporter-otlp-proto-http==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opt_einsum==3.4.0
optree==0.19.0
oras==0.2.42
ordered-set==4.1.0
orjson==3.11.8
oss2==2.19.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==26.0
paho-mqtt==2.1.0
pandas==3.0.2
paramiko==4.0.0
parso==0.8.5
pathable==0.5.0
pdf2docx==0.5.8
pdfrw==0.4
pefile==2023.2.7
pillow==11.3.0
pip==25.3
platformdirs==4.4.0
playwright==1.58.0
portalocker==3.2.0
posthog==7.11.2
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.31.1
proxy_tools==0.1.0
psutil==7.0.0
pure_eval==0.2.3
py-key-value-aio==0.4.4
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycdlib==1.14.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.13.1
pydantic-settings==2.14.0
pydantic_core==2.46.1
pydub==0.25.1
pyee==13.0.1
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.12.1
PyMonCtl==0.92
PyMuPDF==1.26.4
PyNaCl==1.6.2
pynput==1.8.1
pyparsing==3.2.5
PyPDF2==3.0.1
pyperclip==1.11.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-datauri==3.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-engineio==4.13.1
python-frontmatter==1.1.0
python-multipart==0.0.26
python-socketio==5.16.1
python-socks==2.8.1
python-telegram-bot==22.7
pythonnet==3.0.5
pytz==2025.2
pywebview==6.2.1
pywin32==311
pywin32-ctypes==0.2.3
PyWinBox==0.7
PyWinCtl==0.4.1
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.17.1
questionary==2.1.1
qwen-agent==0.0.34
qwen-vl-utils==0.0.14
qwenpaw==1.1.4b1
qwenpaw==1.1.4b1
redis==6.4.0
referencing==0.37.0
regex==2026.4.4
reme_ai==0.3.1.8
reportlab==4.4.3
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
safetensors==0.7.0
scrapling==0.4.6
segno==1.6.6
selenium==4.43.0
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.5
SQLAlchemy==2.0.49
sqlite-vec==0.1.9
sse-starlette==3.3.4
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
tavily-python==0.7.23
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tiktoken==0.12.0
tkinterdnd2==0.4.3
tld==0.13.2
tokenizers==0.22.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==5.6.0
trio==0.33.0
trio-websocket==0.12.2
twilio==9.10.5
Twisted==25.5.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uncalled-for==0.3.1
unpaddedbase64==2.1.0
urllib3==2.6.3
uvicorn==0.45.0
vine==5.1.0
vncdotool==1.3.0
w3lib==2.4.1
watchfiles==1.1.1
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==16.0
wecom-aibot-python-sdk==1.0.2
Werkzeug==3.1.6
wheel==0.46.3
win32_setctime==1.2.0
WMI==1.5.1
wrapt==2.1.1
wsproto==1.3.2
yarl==1.23.0
yaspin==3.4.0
zipp==3.23.1
zope.event==6.0
zope.interface==8.0.1
zstandard==0.25.0
